transform_code -> transform_ent converted
