+++ ts/src/tsc/transform_schema.ts
@@ -399,9 +399,9 @@ function findImportPath(sourceFile: ts.SourceFile) {
 export class TransformSchema implements TransformFile {
   glob = "src/schema/*.ts";
 
-  filter(files: string[]) {
-    return files.filter((f) => f.endsWith("event.ts"));
-  }
+  // filter(files: string[]) {
+  //   return files.filter((f) => f.endsWith("event.ts"));
+  // }
 
   traverseChild(
     sourceFile: ts.SourceFile,
